Add test for ParquetReaderBenchmark
